+++ components/02-elements/form/file-upload/file-upload.js
@@ -5,7 +5,8 @@
 
   components.forEach(component => {
     const input = component.querySelector('.file-upload__input'),
-          name  = component.querySelector('.file-upload__name');
+          name   = component.querySelector('.file-upload__name'),
+          button = component.querySelector('span[role="button"]');
 
     input.addEventListener('change', file => {
       const filePath = file.target.value,
@@ -17,6 +18,13 @@
       else {
         name.textContent = name.dataset.defaultText;
       }
+      name.focus();
     });
-  })
-}());
+
+    button.addEventListener('keypress', event => {
+      if (event.which === 32 || event.which === 13) {
+        input.click();
+      }
+    });
+  });
+})();